[ADD] Tests

+++ src/models/GastoNormal.java
@@ -1,7 +1,7 @@
 package models;
 
 public class GastoNormal extends Gasto {
-    public GastoNormal(Integer monto, String nombre, Expensa expensa) {
+    public GastoNormal(Double monto, String nombre, Expensa expensa) {
         super(monto, nombre, expensa);
     }
 }
